move firebug.sdk APIs shims in a separate file and raise an error on unknown Options
